Switch whitespace linter to iterate over UTF-8 code units.
Keeping the actual/expected sources as `String`s is unnecessary
because we're only scanning them for runs of differing whitespace.
This switches to storing arrays of UTF-8 code units and working
in terms of those arrays and `ArraySlice`s, avoiding all the
overhead of `String` indexing w.r.t. grapheme boundaries.