Use `SourceLocationConverter` for whitespace linter diagnostics.
Previously, we were scanning the input string to compute the line
and column number for diagnostics emitted by the whitespace linter
(and doing this scan for _every_ diagnostic). Since we already
have a `SourceLocationConverter` from the parsed source tree, we
can just use it instead—it's already cached a mapping from UTF-8
offsets to line/column numbers.