Optimize splitting of whitespace runs.
According to Time Profiler, a great deal of time was spent in
the `Collection.split` function, specifically allocating and
deallocating the small arrays that held the newline-delimited
whitespace runs. This has been replaced by a hand-written
sequence that lazily computes the runs and avoids all those
allocations. Even doing two passes over the whitespace (first
to compute the number of runs, then to lazily compute the runs
themselves) is significantly faster than hitting the heap.