Peak memory use optimizations.
Reduce peak memory usage of swift-format by discarding collections once they are no longer needed.
- Remove `Token` arrays from `beforeMap` and `afterMap` while processing the associated token.
- Empty `preVisitedExprs` after visitation finishes for the root expression of the tree.

By the end of a file, these maps can accumulate a meaningful number of tokens such that they significantly impact the peak memory usage.

This admittedly replaces a O(1) lookup operation with a O(N) lookup-and-remove operation, by replacing `subscript` with `removeValue(forKey:)`. In my testing, I found that this had no impact on the runtime. My hypothesis is that removing from these dictionaries keeps them small enough that O(N) runtime isn't a problem.

This set previously contained `ExprSyntax` objects and was maintained for the entire file. Once the visitor has finished with all children of the node that initiated inserting the contextual breaks, it will never visit any of the children again. This set can be emptied at that point to keep it from growing until the visitor finishes with the whole file.

The runtime is unchanged, but this reduces peak memory usage especially for large files.

- Large file (30k LOC): 108.3 MB peak to 72.1 MB peak
- Medium file (10k LOC) : 35.9 MB peak to 24.4 MB peak
- swift-protobuf library (many small files): 77.6 MB peak to 75.6 MB peak