Add portable Darwin build
I use this script for distributing [DrString][], which also depends on `lib_InternalSwiftSyntaxParser.dylib`.

Summary of approach: distribute a copy of `lib_InternalSwiftSyntaxParser.dylib`, modify the main executable's rpath to include the copy's relative path. This ensures the .lib is always compatible with the executable. The resulting .tar.gz can be unpacked into, say, the Homebrew installation folders. And users no longer need the right Xcode release to run `swift-format`.

[DrString]: https://github.com/dduan/DrString/blob/master/Scripts/package-darwin.sh