Group the try keyword with a relevant portion of its expression.
This keeps the try keyword and a portion of the expression that may throw on the same line, when possible. Previously, the group around certain member access exprs would result in frequently breaking after the keyword. That group is replaced by a group that starts at the try keyword now instead.