Add build-only support for the `async` keyword.
Since the parser's behavior is gated on the `-enable-experimental-concurrency`
compiler flag, we unfortunately can't write any tests for this yet as the
in-memory parser will not have the concurrency support enabled. This change
only adds minimal support for `async` in the pretty printer and fixes a
build breakage in `master` caused by the addition of the new keyword.