Support opt-in rules using an `isOptIn` class property.
The codegen pipeline evaluates the new property for rule types that are discovered and uses it to write the correct default value into `RuleRegistry`. No rules are converted to opt-in status as part of this change.

The new property needs to be support overriding in subclasses, so I had to define it as an open class property in the 2 superclasses for rules (`SyntaxFormatRule` and `SyntaxLintRule`).