Add an exception to AlwaysUseLowerCamelCase for test names.
When using descriptive names, e.g. explaining the context and expected outcome in the name, the test names can be quite long and using underscores in the names to organize phrases becomes beneficial. In other situations (i.e. not test cases), it's better to use shorter names that don't require such organization.

This change adds an exception specifically for test cases so that the linter won't complain about underscores in those names.