Specify the type of decl to rename in AlwaysUseLowerCamelCase.
Previously, the diagnostic message said "rename *variable*..." regardless of the actual type of decl. I have added "function", "enum case", "constant", and "identifier" to the list of supported options. This approach is easily scalable to naming other types of decls as well.

I believe this makes the diagnostics more clear and specific, because it was confusing to see a diagnostic that says "rename variable..." that refers to a function's name.