Don't require a returns comment for functions that never return.
It's possible for someone to use a typealias for `Never`, instead of directly using `Never`, but that would be very strange and isn't supported. Any function whose return type is `Never` is ignored when validating returns comments: there won't be a warning to add a comment nor a warning to remove an existing comment. It might be useful to document the fact that the function doesn't return, or that might be trivial and not worth documenting.