Fix and improve parallel mode
Parallel mode was crashing with bad memory access due to data races
accessing `Frontend.configurationLoader`, more specifically its `cache`.

After serializing access (initially with an `NSLock`), I observed that
despite not crashing anymore, the performance was surprisingly *worst*
than in single threaded mode.

That led me down a road of investigating why this was happening, and
after some profiling I discovered that `Rule`'s `nameCache` was the
main source of contention - causing around 30% of total spent time
waiting for locks (`ulock_wait`) on my tests.

After replacing the synchronization mechanism on `nameCache` with a
more efficient `os_unfair_lock_t` (`pthread_mutex_t` in Linux), the
contention dropped significantly, and parallel mode now outperformed
single threaded mode as expected. As a bonus, these changes also
improved single threaded mode performance as well, due to the reduced
overhead of using a lock vs a queue.

I then used the same `Lock` approach to serialize access to
`Frontend.configurationLoader` which increased the performance gap
even further.

After these improvements, I was able to obtain quite significant
performance gains using `Lock`:

- serial (non optimized) vs parallel (optimized): ~5.4x (13.5s vs 74s)
- serial (optimized) vs serial (non optimized): ~1.6x (44s vs 74s)
- serial (optimized) vs parallel (optimized): ~3.2x (13.5s vs 44s)

Sadly, a custom `Lock` implementation is not ideal for `swift-format`
to maintain and Windows support was not provided. As such, The
alternative would be to use `NSLock` which being a part of `Foundation`
is supported on all major platforms.

Using `NSLock` the improvements were not so good, unfortunately:

- serial (non optimized) vs parallel (NSLock): ~1,9x (38s vs 74s)
- serial (NSLock) vs serial (non optimized): ~1,4x (52s vs 74s)
- serial (NSLock) vs parallel (NSLock): ~1,3x (38s vs 52s)

The best solution was then to try and remove all concurrency contention
from `swift-format`. By flattening the `FileToProcess` array and
making the `Rule`'s `nameCache` static, close to optimal performance
should be achievable.

Base rules `SyntaxFormatRule` and `SyntaxLintRule` should be kept
in the `SwiftFormatCore` target to eventually support dynamic
discovery of rules (ditching `generate-pipeline`) and only require
rule implementers to import `SwiftFormatCore`.

On the other hand, any statically generated `ruleName` cache must be in
`SwiftFormatRules` target, because  that's where rule implementations
reside to capture their `ObjectIdentifier` at compile time.

So, in order to have a static rule name cache that's accessible from
`SwiftFormatCore` (especially from `SyntaxFormatRule`), a solution was
to inject a `ruleNameCache` dictionary in `Context`, which is injected
into every `Rule`.

This allows using `generate-pipelines` to a produce a `ruleNameCache`
in `SwiftFormatRules` that's injected on all `Context`s by each of the
pipelines (which is done in `SwiftFormat` target).

While not as ideal as having the cached accessed in a single place
(this approach incurs in the cache being copied for each `Context`), it
achieves a good tradeoff given current constraints). Lets hope the
compiler is smart enough to optimize this somehow.

All things considered, results were quite good:

- serial (non optimized) vs parallel (optimized): ~7x (10.5s vs 74s)
- serial (optimized) vs serial (non optimized): ~1.7x (42.5s vs 74s)
- serial (optimized) vs parallel (optimized): ~4x (10.5s vs 42.5s)

Tests were made on my `MacBookPro16,1` (8-core i9@2.4GHz), on a project
with 2135 Swift files, compiling `swift-format` in Release mode.

## Changes

- Make file preparation serial by calculating all `FileToProcess`
before processing them in parallel.

- Update `Context` to receive a `ruleNameCache` of type
`[ObjectIdentifier: String]` on `init`.

- Update `Context.isRuleEnabled` to receive the `Rule` type instead of
the rule's name, to use the new cache.

- Add new `RuleNameCacheGenerator` to `generate-pipeline` which
outputs a static `ruleNameCache` dictionary to
`RuleNameCache+Generated.swift` file in `SwiftFormatRule` target with
all existing rules.

- Remove the `nameCache` and `nameCacheQueue` `Rule.swift`.