Use the `SwiftOperators` module to fold sequence expressions.
This change deletes the custom folding implementation that was
used previously in swift-format. Folding was previously only performed
before the pretty-printing stage (after the "rules" stage), but now
folding occurs before the folding stage, allowing us to write rules
that are aware of operator precedence.

The exception to the above is that if you use the `format` or `lint`
APIs that take a `SourceFileSyntax`, you must manually fold the
syntax tree first (any `SequenceExpr`s left in the tree will trap).