Fix parallel mode on larger projects
Since the parallel mode requires the count of the files that it's going
to process, the application of `openAndPrepareFile` was not lazy, this
meant that swift-format immediately opened all the files in the project,
which because of macOS's low open file limit quickly results in "too
many open files". This changes the logic to only open the file as
needed instead. This shouldn't be any different from before as far as
file discovery, since we were already resolving everything to fetch the
count.