Revert the use of `generate-pipeline-plugin` on Windows platforms.
This temporarily removes the use of the plugin only on Windows, where it
is not working properly due to a likely bug in SPM. We can keep the
files checked in for now and updated using the
`Scripts/generate-all-pipelines.sh` script, and those files will be
ignored on other platforms (the plugin will still be used there).

Fixes #422.