Workaround new trivia behavior in SwiftSyntax.
As of https://github.com/apple/swift-syntax/pull/985, the parser treats
all trivia following a token up to the next line break as trailing trivia,
including comments. This is a change from the former logic, which treated
such comments as leading trivia of the following token.

A great deal of the logic in swift-format was written in terms of that
behavior, and attempts to rewrite it to use the new behavior instead have
been difficult to get right in all cases. In the interest of expediency,
I'm merging this workaround that shifts the trivia around to restore the
old tree that we expected.

This shouldn't be considered a long-term fix; we should eventually fix
the misbehavior with the new layout and drop the extra mutation.