Fix postfix-`#if` formatting when they come after a closing parenthesis.
A previous change caused the following formatting, which is less than ideal:

```swift
SomeView(
  args
)
  #if os(iOS)
    .someModifier()
  #endif
```

In this case, the `#if` block should be at the same identation level as the
closing parenthesis, similar to how we handle closing braces (`}`).

This change moves the treatment of breaks for postfix-`#if` out of the
visitor for `#if` clauses and instead uses the same contextual break
previsition we use for other member access expressions. The result handles
more cases correctly, and is a simpler, less special-casey implementation.