Move the default `Configuration.init()` into a separate file.
This is meant to make it easier for users forking or building
their own swift-format to change the hardcoded default
configuration.

This also adds a helper for the default test configuration. We
can't let the unit tests rely on the default configuration
because that would make the tests fail when run with a
swift-format binary that has a different default configuration.
This approach lets the default test configuration differ from
the "real" default configuration.