Replace the `ReplaceTrivia` rewriter with direct trivia mutations.
Previously, we were using a `SyntaxRewriter` to replace trivia on
nodes. Historically, this originated from a time (I believe) when
trivia wasn't directly mutable on nodes. The vast majority of these
uses were replacing the extents: the leading trivia of the first
token of a node or the trailing trivia of the last token. This can
be done more easily now by just mutating those properties of the
nodes.

Even in cases where we were doing something slightly more advanced,
it's clearer to do in-place mutation and reconstruction; having
this specialized rewriter no longer held its weight.