Refactor tests.
This PR significantly improves the testing infrastructure by adopting
the emoji markers used by SwiftSyntax's parser tests. This makes it
easier to read and write tests, because the markers indicate in the
string where the finding should be emitted and we don't have to compute
line/column numbers manually.

Additional testing improvements include:

- Tests are no longer stateful w.r.t. collecting findings, so a single
  test can safely call its `assert*` helpers multiple times.
- Tests verify the actual text of the findings, not just their symbolic
  values/constructors. Previously, we weren't actually verifying that
  the *messages* were what we wanted them to be, and this has revealed
  a couple minor issues.
- I've dropped the `XCT*` prefix on our own custom assertions, because
  we shouldn't be using their namespace.

I've added a handful of FIXME comments for things I caught during the
migration that need to be addressed.