Ignore symlinks and hidden (dot) files during `--recursive`.
Note that hidden files will *always* be honored if they are
specified explicitly among the paths on the command line. Symlinks
encountered during recursive traversal *or* passed on the command
line will only be followed if the flag `--follow-symlinks` is
passed. This is meant to avoid symlinks being hidden by shell
expansions; for example, running `swift-format *` where the
current directory contains a symlink and following that symlink
might be surprising to users.

This PR also adds a new test target for code in the `swift-format`
binary in order to test `FileIterator` without factoring the
binary's code out to a separate library target. We should use this
in the future to expand coverage of other functionality like the
frontends and commands.