Don’t set `SWIFT_BUILD_SCRIPT_ENVIRONMENT` when building swift-format
Setting `SWIFT_BUILD_SCRIPT_ENVIRONMENT` entails that we e.g. enable assertions in swift-syntax, even in release mode, which we don’t want since the copy of swift-format that we include in the toolchain is also built using this script and shouldn’t have these swift-syntax’s assertions enabled, to improve performance.