Remove the legacy trivia workaround.
This was added when SwiftSyntax changed its behavior regarding trailing
trivia -- previously it was only whitespace but now it includes all
trivia up to the next newline, including comments (block comments and
end-of-line comments). This workaround rewrote trivia to correspond to
the old layout before processing the syntax tree further so that we
didn't have to overhaul all the rules at the time.

However, it isn't good for long-term maintenance for swift-format to be
written with assumptions about trivia that aren't consistent with how
syntax trees Straight Outta The Parser are formed.