OrderedImports: Fix trivia placement for trailing comments.
Trailing comments were being appended to the next token's leading
trivia instead of staying as trailing trivia; I forgot to make this
adjustment when removing the legacy trivia workaround. Interestingly,
this didn't surface as a problem in either the `OrderedImports`
rule's unit tests or the pretty printer unit tests. Only the
interaction between the two would surface the error.