[AllPublicDeclarationsHaveDocumentation] Fix behavior for `override` methods.
This was mistakenly broken during a refactor and we didn't have test coverage
for it. We didn't have test coverage for it because the tests behave
differently than the rule does when executed in the whole pipeline, so we
need to fix that too. I'm going to do that in a separate PR, because it
might have knock-on effects elsewhere.

Fixes #651.