Modify tests to check the rule and the pipline have the same findings.
Previously we only tested the rule in isolation. This mean we weren't actually checking the pipeline respected the SyntaxVisitorContinueKind. Regardless of what the pipeline did, the tests would respect the continue kind because they run the rule itself.
Now we run the rule in isolation and the pipeline and assert that they have the same findings. This ensures the rule behaves as expected (or atleast as tested) when run in the pipeline.