[OrderedImports] Fix dropped trailing comments on top-level code items.
Also add some validation logic in format rule tests. We already run the
pretty printer afterwards to verify that there aren't any invariants
broken that would cause an assertion failure, but we don't compare the
actual pretty-printed text to the originally transformed tree (because
we don't want the output of those tests to be sensitive to
pretty-printer configuration). What we *can* do, which is still an
improvement, is walk the token sequence and compare the tokens and
trivia in a whitespace-insensitive manner. This makes sure that we
don't move trivia around in a way that the format rule would accept
but that the pretty-printer wouldn't know how to handle.