Migrate the parallel mode to Swift concurrency.
This replaces `Dispatch.concurrentPerform` with a task group. One
benefit of this approach is that we don't need to know the number
of iterations that we want ahead of time. We can start tasks during
the initial iteration over the `FileIterator` instead of eagerly
expanding it first; this should potentially improve behavior for
invocations that cover very large directories recursively,
especially when they're on slower storage like a network drive.