Support for formatting a selection (given as an array of ranges) <https://github.com/apple/swift-format/issues/297>.
The basic idea here is to insert `enableFormatting` and `disableFormatting` tokens into the print stream when we enter or leave the selection. When formatting is enabled, we print out the tokens as usual. When formatting is disabled, we turn off any output until the next `enableFormatting` token. When that token is hit, we write the original source text from the location of the last `disableFormatting` to the current location.

Note that this means that all the APIs need the original source text to be passed in.

A `Selection` is represented as an enum with an `.infinite` case, and a `.ranges` case to indicate either selecting the entire file, or an array of start/end utf-8 offsets. The offset pairs are given with `Range<AbsolutePosition>`, matching the (now common) usage in swift-syntax.

For testing, allow marked text to use `⏩` and `⏪` to deliniate the start/end of a range of a selection.

The command line now takes an `--offsets` option of comma-separated "start:end" pairs to set the selection for formatting.